move the import fast path from IMPORT_NAME to __import__
