+++ pypy/interpreter/pyopcode.py
@@ -1063,7 +1063,6 @@ class __extend__(pyframe.PyFrame):
             return oparg
 
     def IMPORT_NAME(self, nameindex, next_instr):
-        from pypy.module.imp.importing import import_name_fast_path
         space = self.space
         w_modulename = self.getname_w(nameindex)
         w_fromlist = self.popvalue()
@@ -1081,16 +1080,8 @@ class __extend__(pyframe.PyFrame):
             w_locals = space.w_None
         w_globals = self.get_w_globals()
 
-        # the space.w_default_importlib_import attribute is written to in the
-        # startup() method of _frozen_importlib
-        w_default_import = space.w_default_importlib_import
-        if (w_default_import is not None and
-                space.is_w(w_default_import, w_import)):
-            w_obj = import_name_fast_path(space, w_modulename, w_globals,
-                    w_locals, w_fromlist, w_flag)
-        else:
-            w_obj = space.call_function(w_import, w_modulename, w_globals,
-                    w_locals, w_fromlist, w_flag)
+        w_obj = space.call_function(w_import, w_modulename, w_globals,
+                w_locals, w_fromlist, w_flag)
 
         self.pushvalue(w_obj)
