+++ pypy/module/imp/importing.py
@@ -91,7 +91,7 @@ def _readall(space, filename):
 def importhook(space, modulename, w_globals=None, w_locals=None, w_fromlist=None, level=0):
     # A minimal version, that can only import builtin and lib_pypy modules!
     # The actual __import__ is
-    # pypy.module._frozenimportlib.interp_import.import_with_frames_removed
+    # pypy.module._frozenimportlib.interp_import.interp___import__
     assert w_locals is w_globals
     assert level == 0
 
@@ -426,73 +426,6 @@ def load_compiled_module(space, w_modulename, w_mod, cpathname, magic,
     exec_code_module(space, w_mod, code_w, cpathname, cpathname, write_paths)
 
     return w_mod
-
-class FastPathGiveUp(Exception):
-    pass
-
-def _gcd_import(space, name):
-    # check sys.modules, if the module is already there and initialized, we can
-    # use it, otherwise fall back to importlib.__import__
-
-    # NB: we don't get the importing lock here, but CPython has the same fast
-    # path
-    w_modules = space.sys.get('modules')
-    w_module = space.finditem_str(w_modules, name)
-    if w_module is None:
-        raise FastPathGiveUp
-
-    # to check whether a module is initialized, we can ask for
-    # module.__spec__._initializing, which should be False
-    try:
-        w_spec = space.getattr(w_module, space.newtext("__spec__"))
-    except OperationError as e:
-        if not e.match(space, space.w_AttributeError):
-            raise
-        raise FastPathGiveUp
-    try:
-        w_initializing = space.getattr(w_spec, space.newtext("_initializing"))
-    except OperationError as e:
-        if not e.match(space, space.w_AttributeError):
-            raise
-        # we have no mod.__spec__._initializing, so it's probably a builtin
-        # module which we can assume is initialized
-    else:
-        if space.is_true(w_initializing):
-            raise FastPathGiveUp
-    return w_module
-
-def import_name_fast_path(space, w_modulename, w_globals, w_locals, w_fromlist,
-        w_level):
-    level = space.int_w(w_level)
-    if level == 0:
-        # fast path only for absolute imports without a "from" list, for now
-        # fromlist can be supported if we are importing from a module, not a
-        # package. to check that, look for the existence of __path__ attribute
-        # in w_mod
-        try:
-            name = space.text_w(w_modulename)
-            w_mod = _gcd_import(space, name)
-            have_fromlist = space.is_true(w_fromlist)
-            if not have_fromlist:
-                dotindex = name.find(".")
-                if dotindex < 0:
-                    return w_mod
-                return _gcd_import(space, name[:dotindex])
-        except FastPathGiveUp:
-            pass
-        else:
-            assert have_fromlist
-            w_path = space.findattr(w_mod, space.newtext("__path__"))
-            if w_path is not None:
-                # hard case, a package! Call back into importlib
-                w_importlib = space.getbuiltinmodule('_frozen_importlib')
-                return space.call_method(w_importlib, "_handle_fromlist",
-                        w_mod, w_fromlist,
-                        space.w_default_importlib_import)
-            else:
-                return w_mod
-    return space.call_function(space.w_default_importlib_import, w_modulename, w_globals,
-                                w_locals, w_fromlist, w_level)
 
 def get_spec(space, w_module):
     try:
